pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-agent"
version = "1.0.0"
description = "Modular AI Agent with Skill, MCP, and Memory support"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.25.0",
    "openai>=1.20.0",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "pydantic>=2.6.0",
    "redis>=5.0.0",
    "httpx>=0.27.0",
    "tiktoken>=0.6.0",
    "python-dotenv>=1.0.0",
    "structlog>=24.1.0",
    "tenacity>=8.2.0",
    "asyncio-throttle>=1.0.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
vector = ["qdrant-client>=1.8.0", "sentence-transformers>=2.6.0"]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.23.0", "pytest-cov>=5.0.0"]

# ── Knowledge Base optional dependencies ──────────────────────────────────────
kb-docling      = ["docling>=2.5.0"]
kb-chunking     = ["chonkie[sentence]>=0.4"]
kb-elastic      = ["elasticsearch[async]>=8.13.0"]
kb-opensearch   = ["opensearch-py>=2.4.0"]
kb-graph        = ["neo4j>=5.19.0"]
kb-unstructured = ["unstructured[all-docs]>=0.12.0"]
kb-markitdown   = ["markitdown[all]>=0.1.0"]
kb-bge          = ["FlagEmbedding>=1.2.0"]
kb-cohere       = ["cohere>=5.0.0"]
kb-milvus       = ["pymilvus>=2.4.0"]
kb-chroma       = ["chromadb>=0.4.0"]
kb-pdf-fallback = ["pypdf>=4.0.0", "python-docx>=1.1.0", "openpyxl>=3.1.0"]
kb-all = [
    "docling>=2.5.0",
    "chonkie[sentence]>=0.4",
    "elasticsearch[async]>=8.13.0",
    "opensearch-py>=2.4.0",
    "neo4j>=5.19.0",
    "unstructured[all-docs]>=0.12.0",
    "markitdown[all]>=0.1.0",
    "FlagEmbedding>=1.2.0",
    "cohere>=5.0.0",
    "pymilvus>=2.4.0",
    "chromadb>=0.4.0",
    "pypdf>=4.0.0",
    "python-docx>=1.1.0",
    "openpyxl>=3.1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
